move server_pipe to common
